refactor(exercise details): carry out progress to component
