+++ src/modules/exercise/exerciseDetails.module.css
@@ -11,33 +11,6 @@
 	box-sizing: border-box;
 }
 
-.progress {
-	grid-area: progress;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	min-width: 1000px;
-	height: 200px;
-}
-
-.collectionTitle {
-	margin-bottom: 24px;
-}
-
-.statistics {
-	display: flex;
-	justify-content: center;
-	gap: 0px 20%;
-	width: 100%;
-}
-
-.modeContainer {
-	display: flex;
-	align-items: center;
-	flex-direction: column;
-}
-
 .contentCollection {
 	grid-area: contentCollection;
 	display: flex;
@@ -53,24 +26,10 @@
 	width: 70%;
 }
 
-@media (max-width: 344px) {
-	.collectionTitle {
-		font-size: var(--font-size-l);
-	}
-}
-
 @media (max-width: 1024px) {
 	.collectionsContainer {
 		padding: 0px 0px;
 		grid-template-rows: 300px 1fr;
-	}
-	.progress {
-		min-width: 100%;
-		align-content: center;
-	}
-	.modeTitle {
-		font-size: var(--font-size-s);
-		margin-bottom: 15px;
 	}
 	.contentCollection {
 		min-width: 100%;
