+++ src/modules/exercise/services/removeLocalStorage.ts
@@ -1,0 +1,5 @@
+export const removeLocalStorage = (key: string) => {
+	if (localStorage.getItem(key)) {
+		localStorage.removeItem(key);
+	}
+};
